implemented the aws lambda to create/update radar data in the dynamodb database
