+++ aws-backend/.gitignore
@@ -1,1 +1,1 @@
-aws-backend.zip
+*/*.zip
